requirements.txt
Flask
tensorflow
keras
numpy
Pillow
gdown
gunicorn
